% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hotpo1.R
\name{hotpo1}
\alias{hotpo1}
\title{Half or triple plus one}
\usage{
hotpo1(x)
}
\arguments{
\item{x}{an integer}
}
\value{
Half the input, or triple it plus one.
}
\description{
Calculates half the input (if input even) or three times the input plus one (if input odd).
}
\examples{
hotpo1(6)
hotpo1(27)
}

