% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hotpo-all.R
\name{hotpo_hi}
\alias{hotpo_hi}
\title{Maximum value of Collatz sequence}
\usage{
hotpo_hi(x)
}
\arguments{
\item{x}{Starting point for sequence.}
}
\value{
Maximum value reached by Collatz sequence starting at the input value.
}
\description{
Calculates maximum value of entire Collatz sequence starting from input value.
}
\examples{
hotpo_hi(6)
hotpo_hi(27)
}

