% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hotpo-all.R
\name{hotpo_len}
\alias{hotpo_len}
\title{Length of Collatz sequence}
\usage{
hotpo_len(x)
}
\arguments{
\item{x}{Starting point for sequence.}
}
\value{
Length of Collatz sequence starting at the input value.
}
\description{
Calculates length of entire Collatz sequence starting from input value.
}
\examples{
hotpo_len(6)
hotpo_len(27)
}

