% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hotpo-all.R
\name{hotpo_seq}
\alias{hotpo_seq}
\title{Collatz sequence.}
\usage{
hotpo_seq(x)
}
\arguments{
\item{x}{Starting point for sequence.}
}
\value{
The entire Collatz sequence starting at the input value.
}
\description{
Calculates entire Collatz sequence starting from input.
}
\examples{
hotpo_seq(6)
hotpo_seq(27)
}

